+++ ubus.h
@@ -1,14 +1,9 @@
 /*
  * Copyright (C) 2017 John Crispin <blogic@openwrt.org>
  *
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License version 2.1
- * as published by the Free Software Foundation
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
+ * SPDX-License-Identifier: LGPL-2.1
  */
 
-extern void ubus_init(void);
+#pragma once
+
+void ubus_init(void);
